Add New Piece
Chef's favorite game is chess. Today, he invented a new piece and wants to see its strengths.

From a cell (X, Y), the new piece can move to any cell of the chessboard such that its color is different from that of (X, Y).

The new piece is currently located at cell (A, B). Chef wants to calculate the minimum number of steps required to move it to (P, Q).

Note: A chessboard is an 8 x 8 square board, where the cell at the intersection of the i-th row and j-th column is denoted (i,j). Cell (i, j) is colored white if i + j is even and black if i + j is odd.